ACS iOS 2.0.1
The whole app will be supported in both landscape and portrait
orientations as we go forward. Upon touching the “Complaints and
Concerns” button, the user is taken to the ReportFormViewController
which displays the buttons for category selection including “Broken
Street Signs,” “Report Road Damages,” “Report Graffiti,” and “Other
Concerns.” This screen has support for a scrollview. I still need to
add UI constraints to have the UI elements properly align themselves.